# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is an Official Documentation for Go REST API with Fiber Framework",
        "title": "Go REST API with Fiber Framework",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Fatih Rizqon",
            "email": "fatihrizqon@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:3000",
    "basePath": "/",
    "paths": {
...
        "/api/v1/coa-groups": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve all coa group records with pagination",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "COAGroup"
                ],
                "summary": "Get all coa groups",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search keyword",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort column",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort direction (asc, desc)",
                        "name": "order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved all records.",
                        "schema": {
                            "$ref": "#/definitions/response.JSON"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.JSON"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Store a new coa group record",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "COAGroup"
                ],
                "summary": "Create coa group",
                "parameters": [
                    {
                        "description": "COA Group Create Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.COAGroupCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "A new record has been stored.",
                        "schema": {
                            "$ref": "#/definitions/response.JSON"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/response.JSON"
                        }
                    }
                }
            }
        },
        "/api/v1/coa-groups/destroy": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete multiple coa groups by their IDs",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "COAGroup"
                ],
                "summary": "Bulk delete coa groups",
                "parameters": [
                    {
                        "description": "Bulk Action Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.BulkActionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "COAGroup destroyed",
                        "schema": {
                            "$ref": "#/definitions/response.JSON"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/response.JSON"
                        }
                    }
                }
            }
        },
        "/api/v1/coa-groups/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve a single coa group by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "COAGroup"
                ],
                "summary": "Get coa group by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "COA Group ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved selected record.",
                        "schema": {
                            "$ref": "#/definitions/response.JSON"
                        }
                    },
                    "404": {
                        "description": "COA Group not found",
                        "schema": {
                            "$ref": "#/definitions/response.JSON"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update coa group data by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "COAGroup"
                ],
                "summary": "Update coa group",
                "parameters": [
                    {
                        "type": "string",
                        "description": "COA Group ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "COA Group Update Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.COAGroupUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Selected record has been updated.",
                        "schema": {
                            "$ref": "#/definitions/response.JSON"
                        }
                    },
                    "404": {
                        "description": "COA Group not found",
                        "schema": {
                            "$ref": "#/definitions/response.JSON"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Remove a coa group record by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "COAGroup"
                ],
                "summary": "Delete coa group",
                "parameters": [
                    {
                        "type": "string",
                        "description": "COA Group ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Selected record has been deleted.",
                        "schema": {
                            "$ref": "#/definitions/response.JSON"
                        }
                    },
                    "404": {
                        "description": "COA Group not found",
                        "schema": {
                            "$ref": "#/definitions/response.JSON"
                        }
                    }
                }
            }
        },
...
        "/api/v1/dropdown/coa-groups": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve a list of coa groups for dropdown selection",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "COAGroup"
                ],
                "summary": "Get coa group dropdown list",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search keyword",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/response.JSON"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.JSON"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "request.BulkActionRequest": {
            "type": "object",
            "required": [
                "ids"
            ],
            "properties": {
                "ids": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "request.COAGroupCreateRequest": {
            "type": "object",
            "required": [
                "code",
                "name",
                "type"
            ],
            "properties": {
                "category": {
                    "type": "string"
                },
                "code": {
                    "type": "string",
                    "minLength": 1
                },
                "name": {
                    "type": "string",
                    "minLength": 1
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "asset",
                        "liability",
                        "equity",
                        "revenue",
                        "expense"
                    ]
                }
            }
        },
        "request.COAGroupUpdateRequest": {
            "type": "object",
            "required": [
                "code",
                "name",
                "type"
            ],
            "properties": {
                "category": {
                    "type": "string"
                },
                "code": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 1
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 1
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "asset",
                        "liability",
                        "equity",
                        "revenue",
                        "expense"
                    ]
                }
            }
        },
...
        "response.JSON": {
            "type": "object",
            "properties": {
                "data": {},
                "errors": {},
                "message": {
                    "type": "string"
                },
                "meta": {
                    "$ref": "#/definitions/response.Meta"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "response.Links": {
            "type": "object",
            "properties": {
                "current_page": {
                    "type": "string"
                },
                "first_page": {
                    "type": "string"
                },
                "last_page": {
                    "type": "string"
                },
                "next_page": {
                    "type": "string"
                },
                "prev_page": {
                    "type": "string"
                }
            }
        },
        "response.Meta": {
            "type": "object",
            "properties": {
                "info": {
                    "type": "string"
                },
                "links": {
                    "$ref": "#/definitions/response.Links"
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "search": {
                    "type": "string"
                },
                "total_count": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}